+++ src/main/java/com/azure/cosmos/kafka/connect/CosmosDBConfig.java
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
 package com.azure.cosmos.kafka.connect;
 
 import static com.azure.cosmos.kafka.connect.CosmosDBConfig.CosmosClientBuilder.createClient;
@@ -50,7 +53,27 @@ public class CosmosDBConfig extends AbstractConfig {
     private static final String COSMOS_CONTAINER_TOPIC_MAP_DISPLAY = "Topic-Container map";
     private static final String COSMOS_CONTAINER_TOPIC_MAP_DOC =
             "A comma delimited list of Kafka topics mapped to Cosmos containers.\n"
-                    + "For example: topic1#con1,topic2#con2.";    
+                    + "For example: topic1#con1,topic2#con2.";
+
+    public static final String COSMOS_GATEWAY_MODE_ENABLED = "connect.cosmos.connection.gateway.enabled";
+    private static final String COSMOS_GATEWAY_MODE_ENABLED_DOC =
+            "Flag to indicate whether to use gateway mode. By default it is false";
+    private static final boolean DEFAULT_COSMOS_GATEWAY_MODE_ENABLED = false;
+
+    public static final String COSMOS_CONNECTION_SHARING_ENABLED = "connect.cosmos.connection.sharing.enabled";
+    private static final String COSMOS_CONNECTION_SHARING_ENABLED_DOC =
+            "If you have set 'connect.cosmos.connection.gateway.enabled' to true, then this configure will not make any difference. "
+            + "By enabling this it allows connection sharing between instances of cosmos clients on the same jvm.";
+    private static final boolean DEFAULT_COSMOS_CONNECTION_SHARING_ENABLED = false;
+
+    public static final String COSMOS_SINK_BULK_ENABLED_CONF = "connect.cosmos.sink.bulk.enabled";
+    private static final String COSMOS_SINK_BULK_ENABLED_DOC = "Flag to indicate whether Cosmos DB bulk mode is enabled for Sink connector. By default it is true.";
+    private static final boolean DEFAULT_COSMOS_SINK_BULK_ENABLED = true;
+
+    public static final String COSMOS_SINK_MAX_RETRY_COUNT = "connect.cosmos.sink.maxRetryCount";
+    private static final String COSMOS_SINK_MAX_RETRY_COUNT_DOC =
+            "Cosmos DB max retry attempts on write failures for Sink connector. By default, the connector will retry on transient write errors for up to 10 times.";
+    private static final int DEFAULT_COSMOS_SINK_MAX_RETRY_COUNT = 10;
         
     public static final String COSMOS_PROVIDER_NAME_CONF = "connect.cosmos.provider.name";
     private static final String COSMOS_PROVIDER_NAME_DEFAULT = null;
@@ -63,20 +86,28 @@ public class CosmosDBConfig extends AbstractConfig {
     public static final int COSMOS_DATABASE_GROUP_ORDER = 2;
     public static final String COSMOS_CLIENT_USER_AGENT_SUFFIX = "APN/1.0 Microsoft/1.0 KafkaConnect/";
 
-    private String connEndpoint;
-    private String connKey;
-    private String databaseName;
-    private String providerName;
+    private final String connEndpoint;
+    private final String connKey;
+    private final String databaseName;
+    private final String providerName;
+    private final boolean bulkModeEnabled;
+    private final boolean gatewayModeEnabled;
+    private final boolean connectionSharingEnabled;
+    private final int maxRetryCount;
     private TopicContainerMap topicContainerMap = TopicContainerMap.empty();
 
     public CosmosDBConfig(ConfigDef config, Map<String, String> parsedConfig) {
         super(config, parsedConfig);
 
-        connEndpoint = this.getString(COSMOS_CONN_ENDPOINT_CONF);
-        connKey = this.getPassword(COSMOS_CONN_KEY_CONF).value();
-        databaseName = this.getString(COSMOS_DATABASE_NAME_CONF);
-        topicContainerMap = TopicContainerMap.deserialize(this.getString(COSMOS_CONTAINER_TOPIC_MAP_CONF));
-        providerName = this.getString(COSMOS_PROVIDER_NAME_CONF);
+        this.connEndpoint = this.getString(COSMOS_CONN_ENDPOINT_CONF);
+        this.connKey = this.getPassword(COSMOS_CONN_KEY_CONF).value();
+        this.databaseName = this.getString(COSMOS_DATABASE_NAME_CONF);
+        this.topicContainerMap = TopicContainerMap.deserialize(this.getString(COSMOS_CONTAINER_TOPIC_MAP_CONF));
+        this.providerName = this.getString(COSMOS_PROVIDER_NAME_CONF);
+        this.bulkModeEnabled = this.getBoolean(COSMOS_SINK_BULK_ENABLED_CONF);
+        this.maxRetryCount = this.getInt(COSMOS_SINK_MAX_RETRY_COUNT);
+        this.gatewayModeEnabled = this.getBoolean(COSMOS_GATEWAY_MODE_ENABLED);
+        this.connectionSharingEnabled = this.getBoolean(COSMOS_CONNECTION_SHARING_ENABLED);
     }
 
     public CosmosDBConfig(Map<String, String> parsedConfig) {
@@ -125,6 +156,33 @@ public class CosmosDBConfig extends AbstractConfig {
                         "none",
                         Importance.MEDIUM,
                         TOLERANCE_ON_ERROR_DOC
+                ).define(
+                        COSMOS_SINK_BULK_ENABLED_CONF,
+                        Type.BOOLEAN,
+                        DEFAULT_COSMOS_SINK_BULK_ENABLED,
+                        Importance.LOW,
+                        COSMOS_SINK_BULK_ENABLED_DOC
+                )
+                .define(
+                        COSMOS_SINK_MAX_RETRY_COUNT,
+                        Type.INT,
+                        DEFAULT_COSMOS_SINK_MAX_RETRY_COUNT,
+                        Importance.HIGH,
+                        COSMOS_SINK_MAX_RETRY_COUNT_DOC
+                )
+                .define(
+                        COSMOS_GATEWAY_MODE_ENABLED,
+                        Type.BOOLEAN,
+                        DEFAULT_COSMOS_GATEWAY_MODE_ENABLED,
+                        Importance.LOW,
+                        COSMOS_GATEWAY_MODE_ENABLED_DOC
+                )
+                .define(
+                        COSMOS_CONNECTION_SHARING_ENABLED,
+                        Type.BOOLEAN,
+                        DEFAULT_COSMOS_CONNECTION_SHARING_ENABLED,
+                        Importance.LOW,
+                        COSMOS_CONNECTION_SHARING_ENABLED_DOC
                 )
                 .defineInternal(
                         COSMOS_PROVIDER_NAME_CONF,
@@ -186,6 +244,22 @@ public class CosmosDBConfig extends AbstractConfig {
 
     public String getProviderName() {
         return this.providerName;
+    }
+
+    public boolean isBulkModeEnabled() {
+        return this.bulkModeEnabled;
+    }
+
+    public int getMaxRetryCount() {
+        return this.maxRetryCount;
+    }
+
+    public boolean isGatewayModeEnabled() {
+        return gatewayModeEnabled;
+    }
+
+    public boolean isConnectionSharingEnabled() {
+        return this.connectionSharingEnabled;
     }
 
     public static void validateConnection(Map<String, String> connectorConfigs, Map<String, ConfigValue> configValues) {